backend/requirements.txt
docker-compose
ipython
pudb
django
django-cors-headers
djangorestframework
markdown
django-filter
virtualenvwrapper
gunicorn
mysqlclient
django-mysql